Add comments, create the UntilSomeoneDiesDFS, and restructure files
